Final changes before submission

+++ handlers/commenthandlers.py
@@ -8,24 +8,57 @@ from models.comment import Comment
 from globals import blog_key
 
 
-class CreateComment(BlogHandler):
-    def post(self, post_id):
+class CreateEditComment(BlogHandler):
+
+    def get(self, post_id, comment_id):
+        if post_id and comment_id and self.user:
+            key = db.Key.from_path('Comment', int(
+                comment_id), parent=blog_key())
+            c = db.get(key)
+
+            if not c:
+                self.error(404)
+                return
+
+            self.render('editcomment.html', content=c.content)
+
+    def post(self, post_id, comment_id=None):
         """
         Adds a new comment to the post
         """
         content = self.request.get('content')
 
-        if not content:
-            self.redirect('/blog/%s' % post_id)
+        key = db.Key.from_path('Post', int(post_id), parent=blog_key())
+        post = db.get(key)
+
+        if not post:
+            self.error(404)
             return
+
         if not self.user:
             user_name = "Anon"
             # self.redirect('/blog')
         else:
             user_name = self.user.name
 
-        comment = Comment(parent=blog_key(), post_id=post_id,
-                          user_name=user_name, content=content)
+        if not content:
+            self.render("permalink.html", post=post, comments=post.comments,
+                        user=self.user,
+                        error_comment="No text provided for comment")
+            return
+        # Create comment
+        if not comment_id:
+            comment = Comment(parent=blog_key(), post=post,
+                              user_name=user_name, content=content)
+        # Edit comment
+        else:
+            key = db.Key.from_path('Comment', int(comment_id),
+                                   parent=blog_key())
+            comment = db.get(key)
+            comment.content = content
+        if not comment:
+            self.error(404)
+            return
         comment.put()
         sleep(0.1)
         self.redirect('/blog/%s' % post_id)
@@ -38,15 +71,21 @@ class DeleteComment(BlogHandler):
 
     def get(self, post_id, comment_id):
         if post_id and comment_id and self.user:
-            print "YOOOO "
             key = db.Key.from_path('Comment', int(
                 comment_id), parent=blog_key())
             c = db.get(key)
-            if c.user_name == self.user.name:
+
+            if c and c.user_name == self.user.name:
 
                 c.delete()
                 sleep(0.1)
                 self.redirect('/blog/%s' % post_id)
+            else:
+                self.error(404)
+                return
+        else:
+            self.error(404)
+            return
 
 
 class EditComment(BlogHandler):
@@ -54,31 +93,3 @@ class EditComment(BlogHandler):
     Edits Comment
     """
 
-    def get(self, post_id, comment_id):
-        if post_id and comment_id and self.user:
-            key = db.Key.from_path('Comment', int(
-                comment_id), parent=blog_key())
-            c = db.get(key)
-            content = c.content
-            self.render("editcomment.html", content=content)
-
-    def post(self, post_id=None, comment_id=None):
-        """ Handles post request (clicking the Submit button)"""
-
-        if not self.user:
-            self.redirect('/blog')
-
-        content = self.request.get('content')
-
-        if content and self.user:
-            if comment_id:
-                key = db.Key.from_path('Comment', int(
-                    comment_id), parent=blog_key())
-                c = db.get(key)
-                c.content = content
-                c.put()
-                self.redirect('/blog/%s' % str(c.post_id))
-        else:
-            error = "content, please!"
-            self.render(
-                "createoreditpost.html", content=content, error=error)
